+++ src/main/java/com/aaronpmaus/jMath/linearAlgebra/Vector3D.java
@@ -162,7 +162,7 @@ import java.lang.IllegalArgumentException;
     // Calculate Normal of BCD plane
     Vector3D norm2 = (b.subtract(c).crossProduct(d.subtract(c)));
     double angle = norm1.angle(norm2);
-    
+
     // if the dot product of 2 vectors is positive, then the angle between them is between [0,90) or
     // (270,360]. If negative, the angle is between (90,270). Use this fact to determine if the
     // dihedral angle should be negative or positive.
@@ -175,18 +175,14 @@ import java.lang.IllegalArgumentException;
     // to the left of the B->A vector. Otherwise, C->D points to the right from the ABC plane, or
     // falls to the right of the B->A vector.
     double signTest = norm1.dotProduct(d.subtract(c));
-    // if the angle is exactly 90, the dot product is 0 and it is ambiquous which side of the
-    // ABC plane vector C->D falls. In this case, take the angle between the ABC norm and the
-    // C->D vector. If the angle is 0, they are codirectional and C->D falls to the left of B->A.
-    // Otherwise, the angle is 180 and C->D falls to the right of B->A.
-    if(Math.abs(angle - 90.0) < 0.0000000001){
-      double norm1_CD_angle = norm1.angle(d.subtract(c));
-      if(Math.abs(norm1_CD_angle - 180.0) < 0.0000000001){
-        angle *= -1.0;
-      }
+    // if the signTest is 0, then the norm and CD are orthogonal, implying that the dihedral
+    // angle is 180. The ambiguity around 0 (imprecision leading to -/+ 0) will result in the
+    // angle being reported at either 180 or -180, both of which are the same. For consistency,
+    // if the signTest is less than the double precision limit (1*10^-14), return the positive
+    // angle. 
+    if(Math.abs(signTest) < 0.00000000000001){
       return angle;
     }
-    //System.out.println("sign test: " + signTest);
     if(signTest < 0){
       angle *= -1.0;
     }