+++ CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Security
 [None]
 
+## [0.12.3] - 2017-12-08
+### Fixed
+- Bug in Vector3D::calculateDihedralAngle(). It was returning the wrong value
+  if the angle was 90. This has been fixed. It has also been modified for
+  consistency, if the angle is 180, it will return 180 instead of -180.
+
 ## [0.12.2] - 2017-11-02
 ### Added
 - Overloaded methods in Vector3D to return Vector3Ds for the vector operations.